requirements.txt
flask>=3.0.0
pymongo[srv]>=4.6.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
opencv-contrib-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.2.0
pandas>=2.2.0
certifi>=2024.2.2
dnspython>=2.6.1
requests>=2.31.0
gunicorn>=21.2.0
